Fixed a few gcc warnings

+++ src/pedestrians.cpp
@@ -729,7 +729,7 @@ bool pedestrian_t::check_path_blocked(ped_manager_t &ped_mgr, point const &dest,
 		plots_to_test[d] = (cur_plot.contains_pt_xy(test_pts[d]) ? plot : ped_mgr.get_global_plot_id_for_pos(city, test_pts[d]));
 		if (plots_to_test[d] == -1) continue; // no or duplicate plot
 
-		if (d == 0 && plot != plots_to_test[d]) { // update current plot so that VFC, etc. works properly
+		if (d == 0 && (int)plot != plots_to_test[d]) { // update current plot so that VFC, etc. works properly
 			plot = plots_to_test[d];
 			ped_mgr.register_ped_new_plot(*this);
 			if (follow_player) {clear_current_dest();} // clear dest in case it's no longer reachable from this plot (next_plot is not adjacent)
